Links onGetGames to game-start, cleans up pause
Links the onGetGames event handler to the game-start button (so that
it displays unprompted at the start of the game).
Cleans up the logic within the pause button function, and the resume
button (ie "keep playing"), and the gameend state for gameTie and
gameWin, to use fewer selectors/lines of code.